Atom Tools: Created base class for document-based applications Moving some duplicated code to a common base class
